+++ packages/jest-circus/src/legacy-code-todo-rewrite/jestAdapterInit.ts
@@ -54,7 +54,7 @@ export const initialize = async ({
   environment: JestEnvironment;
   runtime: Runtime;
   globalConfig: Config.GlobalConfig;
-  localRequire: <T = unknown>(path: string) => T;
+  localRequire: <T = unknown>(path: string) => Promise<T>;
   testPath: string;
   parentProcess: typeof Process;
   sendMessageToJest?: TestFileEvent;
@@ -112,8 +112,9 @@ export const initialize = async ({
 
   // Jest tests snapshotSerializers in order preceding built-in serializers.
   // Therefore, add in reverse because the last added is the first tested.
-  for (const path of [...config.snapshotSerializers].reverse())
-    addSerializer(localRequire(path));
+  for (const path of [...config.snapshotSerializers].reverse()) {
+    addSerializer(await localRequire(path));
+  }
 
   const snapshotResolver = await buildSnapshotResolver(config, localRequire);
   const snapshotPath = snapshotResolver.resolveSnapshotPath(testPath);